+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/action/model/DirectValueActionItem.java
@@ -129,14 +129,10 @@ public class DirectValueActionItem extends ActionItem implements SimpleValue {
     @Override
     public String toStringNoQuotes(LocalData localData) {
         if (value instanceof Double) {
-            if (Double.compare((double) (Double) value, 0) == 0) {
-                return "0";
-            }
+            return EcmaScript.toString(value);
         }
         if (value instanceof Float) {
-            if (Float.compare((float) (Float) value, 0) == 0) {
-                return "0";
-            }
+            return EcmaScript.toString(value);
         }
         if (value instanceof String) {
             return (String) value;
@@ -155,14 +151,10 @@ public class DirectValueActionItem extends ActionItem implements SimpleValue {
     @Override
     public GraphTextWriter appendToNoQuotes(GraphTextWriter writer, LocalData localData) {
         if (value instanceof Double) {
-            if (Double.compare((double) (Double) value, 0) == 0) {
-                return writer.append("0");
-            }
+            return writer.append(EcmaScript.toString(value));
         }
         if (value instanceof Float) {
-            if (Float.compare((float) (Float) value, 0) == 0) {
-                return writer.append("0");
-            }
+            return writer.append(EcmaScript.toString(value));
         }
         if (value instanceof String) {
             return writer.append((String) value);
@@ -193,17 +185,6 @@ public class DirectValueActionItem extends ActionItem implements SimpleValue {
 
     @Override
     public GraphTextWriter appendTo(GraphTextWriter writer, LocalData localData) {
-        if (value instanceof Double) {
-            if (Double.compare((double) (Double) value, 0) == 0) {
-                return writer.append("0");
-            }
-        }
-        if (value instanceof Float) {
-            if (Float.compare((float) (Float) value, 0) == 0) {
-                return writer.append("0");
-            }
-        }
-
         if (value instanceof String) {
             return writer.append("\"").append(Helper.escapeActionScriptString((String) value)).append("\"");
         }